Moved main menu to ui module
